requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.8.1
b2sdk==2.9.3
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
crispy-bootstrap4==2024.10
cssselect2==0.8.0
Django==5.1.6
django-apscheduler==0.7.0
django-crispy-forms==2.3
django-mathfilters==1.0.0
django-redis==5.4.0
django-storages==1.14.6
django-tinymce==4.1.0
django-widget-tweaks==1.4.12
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
logfury==1.0.1
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.0
redis==6.2.0
reportlab==4.3.1
requests==2.32.3
setuptools==78.1.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.4.0
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
